Add files via upload
I was responsible for the improvement of the screen adaptation of the main page, the design of the first-level paging page for the introduction of Soviet and German military armaments, the design of the second-level paging page for the introduction of Soviet and German armaments, and the screen adaptation of the first-level paging page for Soviet and German military generals. It is worth mentioning that the Soviet and German military generals secondary paging surface will also adopt the Soviet and German military armaments introduction of the headset paging surface program.