Add files via upload
I completed the design of the first level pages of the armaments of the two armies in the Soviet-German War and the design of the headphone pages, and filled 12 pages. I also completed the screen adaptive design of the level 1 page and level 2 page for the generals of the Two armies during the Soviet-German War, and improved the design of the level 2 page for the generals of the Two armies during the Soviet-German War. I also made links between the main page and the paging page, and finally improved the text content of the main page.